requirements.txt
Django>=5.0,<6
djangorestframework>=3.15,<4
# >=5.4: версия без pkg_resources в __init__ (нет предупреждения Setuptools 81+)
djangorestframework-simplejwt>=5.4,<6
django-cors-headers>=4.3,<5
django-nested-admin>=4.0,<5
# Setuptools 81+ помечает pkg_resources устаревшим; линия ниже гасит лишние предупреждения,
# если какая-то зависимость всё ещё тянет pkg_resources.
setuptools>=69,<81
